removed CGFloat from Float extension

+++ TKit/Extensions/Float.swift
@@ -6,20 +6,20 @@
 //  Copyright (c) 2014 trifl. All rights reserved.
 //
 
-import UIKit
+import Foundation
 
-extension CGFloat {
+extension Float {
     // common 
-    var thrice: CGFloat { return self * 3.0 }
-    var twice: CGFloat { return self * 2.0 }
-    var half: CGFloat { return self * 0.5 }
-    var third: CGFloat { return self / 3.0 }
-    var fourth: CGFloat { return self * 0.25 }
-    var sixth: CGFloat { return self / 6.0 }
-    var eighth: CGFloat { return self * 0.125 }
-    var twelfth: CGFloat { return self / 12.0 }
-    var sixteenth: CGFloat { return self * 0.0625 }
+    var thrice: Float { return self * 3.0 }
+    var twice: Float { return self * 2.0 }
+    var half: Float { return self * 0.5 }
+    var third: Float { return self / 3.0 }
+    var fourth: Float { return self * 0.25 }
+    var sixth: Float { return self / 6.0 }
+    var eighth: Float { return self * 0.125 }
+    var twelfth: Float { return self / 12.0 }
+    var sixteenth: Float { return self * 0.0625 }
     
     //
-    var inverse: CGFloat { return 1.0 / self }
+    var inverse: Float { return 1.0 / self }
 }
